refactor(bpa-server): build the BPSec key source in config/, like storage
Address review: everything that validates configuration and produces a
resource from it belongs in config/, as config/storage.rs, config/cla.rs
and config/policy.rs already do. This reverses the direction of the
earlier move into bpsec.rs.

Replace PatternKeySource::from_config with config::bpsec::Config::build,
carrying the JWKS loading, the key/binding validation (oct-only, key_ops
required, no duplicate kids, bindings reference known keys), the 0600
permissions check, and the config-validation tests into config/bpsec.rs.
main.rs and watch_keys call config.build() directly.

bpsec.rs keeps the resource and its runtime behaviour: PatternKeySource,
PatternKeyProvider, SecurityRole, the KeySource/KeyProvider impls, the
hot-reload watcher, and the matching/role-gating tests. The Config struct
stays a plain serde type; only the build() method joins it. No behaviour
change.
